+++ src/streetworks/common/__init__.py
@@ -16,6 +16,7 @@ from .from_datex2 import from_datex2
 from .from_jersey import from_jersey
 from .from_kartverket import from_kartverket
 from .from_mallorca import from_mallorca
+from .from_nsw_livetraffic import from_nsw_livetraffic
 from .from_nvdb import from_nvdb
 from .from_nwb import from_nwb
 from .from_ogc_features import from_ogc_features
@@ -89,4 +90,5 @@ __all__ = [
     "from_kartverket",
     "from_jersey",
     "from_tigerweb",
+    "from_nsw_livetraffic",
 ]